fixed dtb, completed house_list

+++ db/migrate/20220813095246_create_house_lists.rb
@@ -3,7 +3,7 @@ class CreateHouseLists < ActiveRecord::Migration[7.0]
     create_table :house_lists do |t|
       t.string :name
       t.integer :total_price
-      t.float :floor_area
+      t.integer :floor_area
       t.integer :station_distance_time
       t.integer :built_year
       t.integer :floor
